Fire
-Оружие стреляет и наносит урон
(но пока что оружие не двигается в сторону выстрела, + нет прицела)

+++ Assets/Scripts/Data/WeaponData.cs
@@ -6,7 +6,7 @@ public class WeaponData : ScriptableObject
     public GameObject _weaponPrefab;
 
     public string _name; //В дальнейшем я бы это сделал через ScriptableObj
-    public float _damage;
+    public int _damage;
     public float _critDamage; //Например в голову
     public float _durability; //Износ
     public float _attackSpeed;
